test: ✅ Color and Style tests as separate modules
